RW-90: use custom Check, Close and FindInPage icons
