+++ upload-app/src/common/icons/custom-icons/Check.tsx
@@ -1,0 +1,27 @@
+import clsx from 'clsx';
+import { IconProps } from 'common/icons';
+
+export const Check: React.FC<IconProps> = ({
+  overrideColor,
+  pathClassName,
+  svgClassName,
+}) => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    width='600'
+    height='600'
+    viewBox='0 0 600 600'
+    version='1.1'
+    className={svgClassName}
+  >
+    <path
+      d='M 498.852 51.855 C 492.270 54.509, 485.916 59.330, 481.791 64.799 C 479.840 67.384, 420.321 150.050, 349.525 248.500 C 278.729 346.950, 220.272 428.060, 219.622 428.745 C 218.694 429.721, 208.495 418.477, 172.469 376.759 C 108.761 302.985, 112.204 306.718, 104.261 302.808 C 98.080 299.765, 96.779 299.500, 88 299.500 C 79.256 299.500, 77.906 299.772, 71.856 302.750 C 64.030 306.603, 57.885 312.879, 54.160 320.825 C 51.836 325.784, 51.500 327.761, 51.500 336.500 C 51.500 345.916, 51.692 346.850, 54.782 352.500 C 56.906 356.384, 87.192 392.185, 140.644 454 C 186.064 506.525, 223.512 549.605, 223.862 549.732 C 224.891 550.107, 542.205 108.665, 545.485 102.296 C 548.176 97.069, 548.471 95.567, 548.485 87 C 548.498 78.784, 548.141 76.733, 545.840 71.825 C 542.273 64.217, 536.327 57.990, 528.752 53.930 C 523.115 50.909, 521.565 50.550, 513 50.280 C 505.520 50.044, 502.512 50.379, 498.852 51.855'
+      stroke='none'
+      fillRule='evenodd'
+      className={clsx(
+        !overrideColor && 'fill-black dark:fill-white',
+        pathClassName
+      )}
+    />
+  </svg>
+);
