+++ upload-app/src/common/icons/custom-icons/Close.tsx
@@ -1,0 +1,27 @@
+import clsx from 'clsx';
+import { IconProps } from 'common/icons';
+
+export const Close: React.FC<IconProps> = ({
+  overrideColor,
+  pathClassName,
+  svgClassName,
+}) => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    width='600'
+    height='600'
+    viewBox='0 0 600 600'
+    version='1.1'
+    className={svgClassName}
+  >
+    <path
+      d='M 73 52.153 C 65.091 55.261, 55.080 65.469, 52.064 73.500 C 48.860 82.032, 49.391 95.208, 53.257 103.056 C 55.484 107.577, 72.137 124.693, 151.462 203.991 L 246.985 299.481 151.626 394.991 C 42.491 504.297, 50.487 494.965, 50.518 513 C 50.533 521.988, 50.875 524.188, 52.892 528.280 C 56.614 535.831, 63.300 542.530, 70.794 546.219 C 76.911 549.231, 78.240 549.500, 86.979 549.500 C 103.928 549.500, 94.686 557.485, 203.755 448.604 L 300.010 352.514 395.755 448.204 C 473.992 526.397, 492.506 544.407, 497 546.697 C 501.940 549.215, 503.519 549.500, 512.515 549.500 C 521.793 549.500, 523.018 549.261, 529.138 546.250 C 536.970 542.397, 543.112 536.126, 546.840 528.175 C 549.164 523.216, 549.500 521.239, 549.500 512.500 C 549.500 503.520, 549.214 501.939, 546.697 497 C 544.407 492.506, 526.397 473.992, 448.204 395.755 L 352.514 300.010 448.604 203.755 C 557.485 94.686, 549.500 103.928, 549.500 86.979 C 549.500 78.240, 549.231 76.911, 546.219 70.794 C 542.530 63.300, 535.831 56.614, 528.280 52.892 C 524.188 50.875, 521.988 50.533, 513 50.518 C 494.965 50.487, 504.297 42.491, 394.991 151.626 L 299.481 246.985 203.991 151.462 C 124.501 71.945, 107.584 55.487, 103.033 53.246 C 95.365 49.469, 81.146 48.951, 73 52.153'
+      stroke='none'
+      fillRule='evenodd'
+      className={clsx(
+        !overrideColor && 'fill-black dark:fill-white',
+        pathClassName
+      )}
+    />
+  </svg>
+);
