+++ upload-app/src/common/icons/custom-icons/FindInPage.tsx
@@ -1,0 +1,27 @@
+import clsx from 'clsx';
+import { IconProps } from 'common/icons';
+
+export const FindInPage: React.FC<IconProps> = ({
+  overrideColor,
+  pathClassName,
+  svgClassName,
+}) => (
+  <svg
+    xmlns='http://www.w3.org/2000/svg'
+    width='600'
+    height='600'
+    viewBox='0 0 600 600'
+    version='1.1'
+    className={svgClassName}
+  >
+    <path
+      d='M 145.124 51.578 C 125.355 56.895, 109.131 73.437, 103.997 93.511 C 102.034 101.188, 102 104.688, 102 300 C 102 495.312, 102.034 498.812, 103.997 506.489 C 107.995 522.121, 118.791 535.992, 132.592 543.226 C 146.391 550.458, 135.094 550, 299.500 550 C 429.673 550, 448.295 549.814, 453.498 548.458 C 456.753 547.611, 459.580 546.753, 459.781 546.552 C 460.373 545.960, 361.031 429, 359.935 429 C 359.384 429, 356.727 430.118, 354.030 431.485 C 346.574 435.263, 337.504 438.473, 327 441.050 C 319.079 442.993, 314.758 443.391, 301 443.440 C 290.723 443.477, 281.860 442.977, 277.500 442.113 C 227.310 432.171, 188.829 393.690, 178.887 343.500 C 178.065 339.353, 177.500 330.182, 177.500 321 C 177.500 304.193, 178.599 296.670, 183.071 282.861 C 206.273 211.216, 288.189 177.291, 355.696 211.370 C 389.628 228.500, 413.562 260.378, 421.113 298.500 C 422.935 307.696, 422.937 334.286, 421.117 343.500 C 417.131 363.682, 407.734 383.540, 394.727 399.268 C 391.271 403.446, 390.952 404.247, 392.142 405.749 C 394.202 408.350, 487.073 518.297, 488.548 519.881 C 489.620 521.032, 490.264 520.412, 492.424 516.150 C 498.194 504.763, 497.969 511.454, 497.985 350.859 L 498 203.218 425.418 126.609 L 352.836 50 251.668 50.066 C 166.716 50.121, 149.638 50.363, 145.124 51.578 M 286.500 237.413 C 253.482 242.912, 227.085 266.454, 217.764 298.715 C 214.513 309.965, 214.515 332.041, 217.768 343.299 C 226.507 373.549, 250.486 396.337, 281.123 403.509 C 291.706 405.986, 309.112 405.784, 319.500 403.063 C 362.178 391.886, 389.813 350.471, 382.932 308 C 381.420 298.666, 377.528 287.643, 372.694 279 C 367.481 269.677, 352.095 253.787, 343.500 248.847 C 325.323 238.402, 305.019 234.329, 286.500 237.413'
+      stroke='none'
+      fillRule='evenodd'
+      className={clsx(
+        !overrideColor && 'fill-black dark:fill-white',
+        pathClassName
+      )}
+    />
+  </svg>
+);
