+++ upload-app/src/common/icons/custom-icons/index.ts
@@ -1,6 +1,9 @@
 export * from './Call';
 export * from './Camera';
 export * from './CameraLine';
+export * from './Check';
+export * from './Close';
+export * from './FindInPage';
 export * from './Forest';
 export * from './ForestLine';
 export * from './Groups';
