--- upload-app/src/common/icons/material-symbols/Close.tsx
@@ -1,24 +1,0 @@
-import clsx from 'clsx';
-import { IconProps } from 'common/icons';
-
-export const Close: React.FC<IconProps> = ({
-  overrideColor,
-  pathClassName,
-  svgClassName,
-}) => (
-  <svg
-    xmlns='http://www.w3.org/2000/svg'
-    height='24'
-    viewBox='0 -960 960 960'
-    width='24'
-    className={svgClassName}
-  >
-    <path
-      className={clsx(
-        !overrideColor && 'fill-black dark:fill-white',
-        pathClassName
-      )}
-      d='m256-200-56-56 224-224-224-224 56-56 224 224 224-224 56 56-224 224 224 224-56 56-224-224-224 224Z'
-    />
-  </svg>
-);
