+++ upload-app/src/common/icons/material-symbols/index.ts
@@ -1,9 +1,6 @@
-export * from './Close';
 export * from './CloudOff';
-export * from './Done';
 export * from './ExpandLess';
 export * from './ExpandMore';
-export * from './FindInPage';
 export * from './ImageLine';
 export * from './InstallMobile';
 export * from './Repeat';
