+++ upload-app/src/modules/new-plant/components/CaptureModal.tsx
@@ -6,7 +6,7 @@ import {
   LoaderBox,
   Section,
 } from 'common/components';
-import { CameraLine, Done, LocationOn, Repeat } from 'common/icons';
+import { CameraLine, Check, LocationOn, Repeat } from 'common/icons';
 import { useFmtMsg } from 'modules/intl';
 import { Layout } from 'modules/layout';
 import { useLogLocationFailed, useLogLocationSucceeded } from 'modules/logging';
@@ -96,7 +96,7 @@ export const CaptureModal: React.FC = () => {
                     {fmtMsg('retake')}
                   </Button>
                   <Button
-                    Icon={Done}
+                    Icon={Check}
                     onClick={() => {
                       store.setImage(store.tmpImage);
                       store.closeCapture();
